Breaking up util: Move custom error to shell package
